+++ StabilityMatrix.Avalonia/Views/InferencePage.axaml.cs
@@ -92,6 +92,11 @@ public partial class InferencePage : UserControlBase
         (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.Krea2ImageEdit);
     }
 
+    private void AddTabMenu_AnimaImageEdit_OnClick(object? sender, RoutedEventArgs e)
+    {
+        (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.AnimaImageEdit);
+    }
+
     private void AddTabMenu_LtxvTextToVideo_OnClick(object? sender, RoutedEventArgs e)
     {
         (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.LtxvTextToVideo);