+++ StabilityMatrix.Core/Models/Api/Comfy/NodeTypes/NodeConnections.cs
@@ -51,3 +51,5 @@ public class SamModelNodeConnection : NodeConnectionBase;
 public class GuiderNodeConnection : NodeConnectionBase;
 
 public class NoiseNodeConnection : NodeConnectionBase;
+
+public class AnimaIpAdapterNodeConnection : NodeConnectionBase;